third_party/WebKit/LayoutTests/http/tests/inspector/tracing/websocket/timeline-websocket-event.html

<html>
<head>
<script src="../../inspector-test.js"></script>
<script src="../../timeline-test.js"></script>
<script>

function performActions()
{
    var ws = new WebSocket("ws://127.0.0.1:8880/simple");
    return new Promise((fulfill) => ws.onclose = fulfill);
}

function test()
{
    InspectorTest.invokeAsyncWithTimeline("performActions", finish);

    function finish()
    {
        InspectorTest.printTimelineRecordsWithDetails("WebSocketCreate");
        InspectorTest.printTimelineRecordsWithDetails("WebSocketSendHandshakeRequest");
        InspectorTest.printTimelineRecordsWithDetails("WebSocketReceiveHandshakeResponse");
        InspectorTest.printTimelineRecordsWithDetails("WebSocketDestroy");
        InspectorTest.completeTest();
    }
}

if (!window.testRunner)
    setTimeout(performActions, 2000);

</script>
</head>

<body onload="runTest()">
<p>
Tests the Timeline events for WebSocket
</p>

</body>
</html>
