third_party/WebKit/LayoutTests/http/tests/inspector/tracing/websocket/timeline-websocket-event-expected.txt

Tests the Timeline events for WebSocket

WebSocketCreate Properties:
{
    data : {
        frame : <string>
        identifier : <number>
        stackTrace : <object>
        url : ws://127.0.0.1:8880/simple
    }
    endTime : <number>
    frameId : <string>
    stackTrace : <object>
    startTime : <number>
    thread : <string>
    type : "WebSocketCreate"
}
WebSocketSendHandshakeRequest Properties:
{
    data : {
        frame : <string>
        identifier : <number>
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    thread : <string>
    type : "WebSocketSendHandshakeRequest"
}
WebSocketReceiveHandshakeResponse Properties:
{
    data : {
        frame : <string>
        identifier : <number>
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    thread : <string>
    type : "WebSocketReceiveHandshakeResponse"
}
WebSocketDestroy Properties:
{
    data : {
        frame : <string>
        identifier : <number>
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    thread : <string>
    type : "WebSocketDestroy"
}
